rr_graph: Reorganize code to check rr_graph.xml so tests work under CI.
