+++ utils/lib/rr_graph/cyclic_to_acyclic.py
@@ -156,8 +156,8 @@ def cyclic_to_acyclic(mapping):
     >>> mapping2 = {0: [1], 1: [2], 2: [3, 5], 3: [6, 9],
     ...             4: [1], 5: [4], 6: [7], 7: [8, 11],
     ...             8: [], 9: [10], 10: [6], 11: [10]}
-    >>> cyclic_to_acyclic(mapping2) == ({0: {1}, 1: {3}, 3: {6, 9}, 6: {8}, 8: set(), 9: {6} }
-    ...                                 {1: 1, 2: 1: 4:1, 5: 1, 6: 5, 7: 6, 11: 6, 10: 6})
+    >>> cyclic_to_acyclic(mapping2) == ({0: {1}, 1: {3}, 3: {6, 9}, 6: {8}, 8: set(), 9: {6} },
+    ...                                 {1: 1, 2: 1, 4: 1, 5: 1, 6: 6, 7: 6, 11: 6, 10: 6})
     True
 
     0 -> 1 -> 2 -> 3
@@ -171,7 +171,7 @@ def cyclic_to_acyclic(mapping):
 
     >>> mapping = {0: [1], 1: [2], 2: [3, 5], 3: [], 4: [1], 5: [4], 6: [7], 7: []}
     >>> cyclic_to_acyclic(mapping) == ({0: {1}, 1: {3}, 3: set(), 6: {7}, 7: set()},
-    ...                                {1: 1, 2: 1, 4: 1 5: 1})
+    ...                                {1: 1, 2: 1, 4: 1, 5: 1})
     True
 
     '''
